I think i have serialization making sense, trying cursor to debug compilation

+++ src/packages/structures/blockChain/transaction.h
@@ -47,7 +47,7 @@ void TW_Transaction_destroy(TW_Transaction* tx);
 size_t TW_Transaction_get_size(const TW_Transaction* tx);
 void TW_Transaction_hash(TW_Transaction* tx, unsigned char* hash_out);
 
-size_t TW_Transaction_serialize(TW_Transaction* tx, unsigned char** out_buffer);
+int TW_Transaction_serialize(TW_Transaction* tx, unsigned char** out_buffer);
 TW_Transaction* TW_Transaction_deserialize(const unsigned char* buffer, size_t buffer_size);
 
 void TW_Transaction_add_signature(TW_Transaction* txn);
